publish-ready/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
openpyxl==3.1.5
pillow==12.2.0
python-multipart==0.0.26
jinja2==3.1.6
